Afina funcionamiento general

+++ components/consumo_endpoints/consumo_endpoints.h
@@ -4,6 +4,6 @@
 #include "esp_err.h"
 
 void obtener_confirmacion_dosificador(int *confirmacion);
-void enviar_consumo_gato(int consumo);
+void enviar_consumo_gato(float consumo);
 
 #endif // CONSUMODATA_H
